requirements.txt
langgraph
langchain
langchain-core
langchain-community
pydantic
python-dotenv
langchain-google-genai
langchain-groq
duckduckgo-search
uvicorn
fastapi
streamlit